Set Output Function Implementation
Implements set output function to set which pin outputs the signal.

Note that I used an addition operation instead of a bitwise OR
operation. This is because MSP430 community coding convention prefers
addition operations over bitwise OR operations. Otherwise, they should
functionally be the same.

If anyone with an MSP430 is willing to test which one is faster, please
feel free to do so and show me the results so I can change the
appropriate code. For now, I will stick to the community convention.